requirements.txt
beautifulsoup4
Django
djangorestframework
pyppeteer
requests
